Refactor: 변경된 타입 경로 수정
