+++ src/components/AuthPage/EmailInput/index.tsx
@@ -1,6 +1,6 @@
 import { Input } from '@/components/common/Input';
 import { PLACEHOLDERS } from '@/constants/Placeholders';
-import { AuthInputProps } from '@/types/AuthType';
+import { AuthInputProps } from '@/types/Auth/AuthInputProps';
 import { emailValidation } from '@/utils/authValidation';
 
 export const EmailInput = ({ register, error }: AuthInputProps) => {
