+++ src/utils/authValidation.ts
@@ -1,5 +1,5 @@
 import { UseFormGetValues } from 'react-hook-form';
-import { AuthDataType } from '@/types/AuthType';
+import { AuthDataRequest } from '@/types/Auth/AuthDataRequest';
 
 export const nameValidation = {
   required: '이름을 입력해주세요.',
@@ -20,13 +20,13 @@ export const emailValidation = {
 export const passwordValidation = {
   required: '비밀번호를 입력해주세요.',
   minLength: {
-    value: 6,
-    message: '비밀번호는 최소 6자 이상이어야 합니다.',
+    value: 4,
+    message: '비밀번호는 최소 4자 이상이어야 합니다.',
   },
 };
 
 export interface PasswordChkValidationProps {
-  getValues?: UseFormGetValues<AuthDataType>;
+  getValues?: UseFormGetValues<AuthDataRequest>;
 }
 
 export const passwordChkValidation = ({
